agregando el segundo reducer

+++ src/components/Producto.jsx
@@ -1,26 +1,53 @@
 import React from 'react';
-import { Link } from 'react-router-dom';
+import { Link, useHistory } from 'react-router-dom';
 import { useDispatch } from 'react-redux';
-import { eliminarProducto } from '../actions/product';
+import { eliminarProducto, editarProducto } from '../actions/product';
+import Swal from 'sweetalert2';
 
 export const Producto = ({producto}) => {
-    const {nombre, precio, id} = producto
+    
+    const {nombre, precio, id} = producto;
     const dispatch = useDispatch();
+    const history = useHistory();
 
     const handleDelete = (id) => {
         // preguntar al usuario
+        Swal.fire({
+            title: 'Quieres borrar este producto?',
+            text: "No podras recuperar este archivo!!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Si',
+            cancelButtonText: 'No'
+          }).then((result) => {
+            if (result.value) {
+                // pasarlo al action
+                dispatch(eliminarProducto(id));
 
-        // pasarlo al action
-        dispatch(eliminarProducto(id))
+            }
+          })
+        
     }
+    // funcion que redirige de forma programada
+    const redireccionarEdicion = producto => {
+        dispatch(editarProducto(producto));
+        history.push(`/products/edit/${producto.id}`)
+    }
+
+
     return (
         <tr>
             <td>{nombre}</td>
             <td><span className='font-wight-bold'>$ {precio}</span></td>
             <td className='acciones'>
-                <Link to={`/products/edit/${id}`} className='btn btn-primary mr-2'>
+                <button
+                type='button'
+                onClick={() => redireccionarEdicion(producto)} 
+                className='btn btn-primary mr-2'>
                     Editar
-                </Link>
+                </button>
                 <button
                     type="button"
                     className="btn btn-danger"
